Change the variable
I change the variable "int" to "float"

+++ Python_Caculator.py
@@ -1,11 +1,11 @@
 # 1- input first numebr.
-first_numebr = int(input("enter first number: "))
+first_numebr = float(input("enter first number: "))
 
 # 2- input operation type.
 operation = input("enter operation type: ")
 
 # 3- input second number.
-second_numebr = int(input("enter second number: "))
+second_numebr = float(input("enter second number: "))
 
 # 4- print the result
 if operation == "+":
